refactor(sideboard_card_selector): split into mixin package
Move widgets/panels/sideboard_card_selector.py into a package: frame.py
builds the scrollable card grid, properties.py holds the getters and
translate helper, and handlers.py owns the +/−/↓/↑ button callbacks and
display refreshers.
