+++ widgets/panels/sideboard_card_selector/__init__.py
@@ -1,0 +1,7 @@
+"""Sideboard card selector panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.sideboard_card_selector.frame import SideboardCardSelector
+
+__all__ = ["SideboardCardSelector"]
